Update solidus_support
A newer version of solidus_support has a method to help us handle full
names inside of addresses, specifically if it is using the full_name
field or the name field.

When I did the upgrade, a whole bunch of other test failures happened
and they all had to do with the auto-loader getting confused with the
namespaces it should be finding things in (I think). So I'm also being
more explicit with the namespaces so they are definitely found where
they should be.